u26

+++ ejemplos/u26/aplicacion/app.py
@@ -58,6 +58,8 @@ def categorias_edit(id):
 @app.route('/categorias/<id>/delete', methods=["get","post"])
 def categorias_delete(id):
 	cat=Categorias.query.get(id)
+	if cat is None:
+		abort(404)
 	form=formSINO()
 	if form.validate_on_submit():
 		if form.si.data:
@@ -87,6 +89,7 @@ def articulos_new():
 	else:
 		return render_template("articulos_new.html",form=form)
 
+@app.route('/articulos/<id>/edit', methods=["get","post"])
 def articulos_edit(id):
 	art=Articulos.query.get(id)
 	if art is None:
@@ -117,9 +120,13 @@ def articulos_edit(id):
 @app.route('/articulos/<id>/delete', methods=["get","post"])
 def articulos_delete(id):
 	art=Articulos.query.get(id)
+	if art is None:
+		abort(404)
+
 	form=formSINO()
 	if form.validate_on_submit():
 		if form.si.data:
+			os.remove(app.root_path+"/static/upload/"+art.image)
 			db.session.delete(art)
 			db.session.commit()
 		return redirect(url_for("inicio"))
